move logic to lenscode
