+++ lens/src/main/java/nl/bvkatwijk/lens/processor/Code.java
@@ -4,8 +4,14 @@ import io.vavr.collection.List;
 import io.vavr.collection.Vector;
 import nl.bvkatwijk.lens.Const;
 
+import javax.lang.model.element.Element;
 import javax.lang.model.element.RecordComponentElement;
+import javax.lang.model.type.TypeMirror;
+import java.util.regex.Pattern;
 
+/**
+ * Code generation for generic java
+ */
 public class Code {
     public static final String PSF = "public static final ";
 
@@ -25,14 +31,6 @@ public class Code {
 
     static String importStatement(String qualifiedTypeName) {
         return "import " + qualifiedTypeName + ";";
-    }
-
-    static String iLens(String from, String to) {
-        return Const.ILENS + "<" + params(from, to) + ">";
-    }
-
-    static String iLens(String name) {
-        return iLens(Const.PARAM_SOURCE_TYPE, name);
     }
 
     // todo not very elegant
@@ -58,5 +56,29 @@ public class Code {
 
     static String ret(String arg) {
         return "return " + arg + ";";
+    }
+
+    static String capitalize(String string) {
+        return Pattern.compile("^.")
+            .matcher(string)
+            .replaceFirst(m -> m.group().toUpperCase());
+    }
+
+    static String typeName(Element it) {
+        TypeMirror type = it.asType();
+        return switch (type.getKind()) {
+            case BOOLEAN -> Boolean.class.getName();
+            case BYTE -> Byte.class.getName();
+            case SHORT -> Short.class.getName();
+            case INT -> Integer.class.getName();
+            case LONG -> Long.class.getName();
+            case CHAR -> Character.class.getName();
+            case FLOAT -> Float.class.getName();
+            case DOUBLE -> Double.class.getName();
+            case VOID -> Void.class.getName(); // Does it make sense to support Void type?
+            case DECLARED -> type.toString();
+            case OTHER, NONE, MODULE, INTERSECTION, UNION, EXECUTABLE, PACKAGE, WILDCARD, TYPEVAR, ERROR, ARRAY, NULL ->
+                throw new IllegalArgumentException("Type " + it + " (" + it.getKind() + " " + type.getKind() + ") not yet supported.");
+        };
     }
 }
