+++ lens/src/main/java/nl/bvkatwijk/lens/processor/LensProcessor.java
@@ -5,8 +5,6 @@ import io.vavr.collection.List;
 import lombok.SneakyThrows;
 import nl.bvkatwijk.lens.Const;
 import nl.bvkatwijk.lens.Lenses;
-import nl.bvkatwijk.lens.kind.ILens;
-import nl.bvkatwijk.lens.kind.Lens;
 
 import javax.annotation.processing.AbstractProcessor;
 import javax.annotation.processing.RoundEnvironment;
@@ -18,7 +16,6 @@ import javax.lang.model.type.DeclaredType;
 import javax.lang.model.type.TypeMirror;
 import java.io.IOException;
 import java.util.Set;
-import java.util.regex.Pattern;
 
 @SupportedSourceVersion(SourceVersion.RELEASE_23)
 @SupportedAnnotationTypes(Const.LENS_ANNOTATION_QUALIFIED)
@@ -44,45 +41,53 @@ public class LensProcessor extends AbstractProcessor {
             .map(RecordComponentElement.class::cast)
             .toList());
 
+        var lens = new DetectedLens(name, fields);
+
         writeSourceFile(Const.PACK, name, String.join(
             "\n",
             List.of("package " + Const.PACK + ";", "")
-                .appendAll(imports(List.of(element)))
+                .appendAll(LensCode.imports(List.of(element)))
                 .append("")
-                .append("public record " + name + Const.LENS + "<" + Const.PARAM_SOURCE_TYPE + ">(" + Code.iLens(name) + " inner) implements " + Code.iLens(
+                .append("public record " + name + Const.LENS + "<" + Const.PARAM_SOURCE_TYPE + ">(" + LensCode.iLens(name) + " inner) implements " + LensCode.iLens(
                     name) + " {")
-                .append(rootLens(name))
-                .appendAll(lensConstants(fields, name))
+                .append(LensCode.rootLens(name))
+                .appendAll(LensCode.lensConstants(fields, name))
                 .appendAll(lensMethods(fields))
-                .appendAll(innerDelegation(name))
+                .appendAll(LensCode.innerDelegation(name))
                 .append("}")
                 .toJavaList()));
     }
 
-    private String rootLens(String name) {
-        return Code.indent(Code.PSF + name + Const.LENS + "<" + name + "> " + Const.ROOT_LENS_NAME + " = new " + name + Const.LENS + "<>(" + Const.BASE_LENS + ".identity());");
+    record DetectedLens(String name, List<RecordComponentElement> fields) {
+//        public String source() {
+//            return List.of("package " + Const.PACK + ";", "")
+//                .appendAll(imports(List.of(element)))
+//                .append("")
+//                .append("public record " + name + Const.LENS + "<" + Const.PARAM_SOURCE_TYPE + ">(" + Code.iLens(name) + " inner) implements " + Code.iLens(
+//                    name) + " {")
+//                .append(rootLens(name))
+//                .appendAll(lensConstants(fields, name))
+//                .appendAll(lensMethods(fields))
+//                .appendAll(innerDelegation(name))
+//                .append("}")
+//                .toJavaList());
+//        }
     }
 
-    private List<String> lensConstants(List<RecordComponentElement> fields, String name) {
-        return fields
-            .map(it -> lensConstant(name, Code.fieldName(it), typeName(it)))
-            .map(Code::indent)
-            .toList();
+    static Iterable<String> lensMethods(List<RecordComponentElement> fields) {
+        return fields.flatMap(LensProcessor::lensMethod);
     }
 
-    private Iterable<String> lensMethods(List<RecordComponentElement> fields) {
-        return fields.flatMap(this::lensMethod);
-    }
-
-    private Iterable<String> lensMethod(RecordComponentElement element) {
-        return qualifiedLens(element).lensMethod();
+    static Iterable<String> lensMethod(RecordComponentElement element) {
+        return qualifiedLens(element)
+            .lensMethod();
     }
 
     private static FieldLens qualifiedLens(RecordComponentElement field) {
         TypeMirror type = field.asType();
         return switch (type.getKind()) {
             case BOOLEAN, BYTE, SHORT, INT, LONG, CHAR, FLOAT, DOUBLE, VOID ->
-                new FieldLens(typeName(field), LensKind.PRIMITIVE, field);
+                new FieldLens(Code.typeName(field), LensKind.PRIMITIVE, field);
             case DECLARED -> declared(field);
             case OTHER, NONE, MODULE, INTERSECTION, UNION, EXECUTABLE, PACKAGE, WILDCARD, TYPEVAR, ERROR, ARRAY, NULL ->
                 throw new IllegalArgumentException("Type " + field + " (" + field.getKind() + " " + type.getKind() + ") not yet supported.");
@@ -94,74 +99,7 @@ public class LensProcessor extends AbstractProcessor {
             .map(AnnotationMirror::toString)
             .contains("@" + Lenses.class.getName())
             ? new FieldLens("unused?", LensKind.LENSED, field)
-            : new FieldLens(typeName(field), LensKind.OTHER, field);
-    }
-
-    static String lensName(RecordComponentElement field) {
-        return lensName(Code.fieldName(field));
-    }
-
-    private Iterable<String> innerDelegation(String typeName) {
-        return Code.indent(List.of(
-            "",
-            "public java.util.function.BiFunction<" + Code.params(
-                Const.PARAM_SOURCE_TYPE,
-                typeName,
-                Const.PARAM_SOURCE_TYPE) + "> with() {",
-            Code.indent(Code.ret("inner.with()")),
-            "}",
-            "",
-            "public java.util.function.Function<" + Code.params(Const.PARAM_SOURCE_TYPE, typeName) + "> get() {",
-            Code.indent(Code.ret("inner.get()")),
-            "}"
-        ));
-    }
-
-    private Iterable<String> imports(List<? extends Element> fields) {
-        return fields
-            .map(LensProcessor::typeName)
-            .map(Code::removeGenerics)
-            .append(ILens.class.getName())
-            .append(Lens.class.getName())
-            .map(Code::importStatement);
-    }
-
-    private static String typeName(Element it) {
-        TypeMirror type = it.asType();
-        return switch (type.getKind()) {
-            case BOOLEAN -> Boolean.class.getName();
-            case BYTE -> Byte.class.getName();
-            case SHORT -> Short.class.getName();
-            case INT -> Integer.class.getName();
-            case LONG -> Long.class.getName();
-            case CHAR -> Character.class.getName();
-            case FLOAT -> Float.class.getName();
-            case DOUBLE -> Double.class.getName();
-            case VOID -> Void.class.getName(); // Does it make sense to support Void type?
-            case DECLARED -> type.toString();
-            case OTHER, NONE, MODULE, INTERSECTION, UNION, EXECUTABLE, PACKAGE, WILDCARD, TYPEVAR, ERROR, ARRAY, NULL ->
-                throw new IllegalArgumentException("Type " + it + " (" + it.getKind() + " " + type.getKind() + ") not yet supported.");
-        };
-    }
-
-    String lensConstant(String record, String field, String fieldType) {
-        return Code.PSF + Code.iLens(record, fieldType) + " " + lensName(field) + " = new " + Const.BASE_LENS + "<>("
-            + Code.params(Code.reference(record, field), Code.reference(record, witherName(field)))
-            + ");";
-    }
-
-    private static String lensName(String field) {
-        return field.toUpperCase();
-    }
-
-    String witherName(String fieldName) {
-        return "with" + capitalize(fieldName);
-    }
-
-    static String capitalize(String string) {
-        return Pattern.compile("^.")
-            .matcher(string)
-            .replaceFirst(m -> m.group().toUpperCase());
+            : new FieldLens(Code.typeName(field), LensKind.OTHER, field);
     }
 
     private void writeSourceFile(String pack, String name, String content) throws IOException {
